Servers a,e,s ok

+++ a/RemoteScripts/check-inner-availability.sh
@@ -1,5 +1,10 @@
-#! /bin/bash -x
-# Check availability of a server from within its Docker container
+#! /bin/bash
+# Check availability of some servers from within its Docker container
+# Usage:  check-inner-availability.sh a.paracamplus.net e.paracamplus.net ...
+#    The arguments are the hostname to request.
+# This script is parameterized via two variables:
+#    URL       the URL to get (by default http://127.0.0.1/
+#    FRAGMENT  the fragment of text that must be present in the response.
 
 HOSTNAME=${HOSTNAME:-no.host.name}
 source ${0%/*}/$HOSTNAME.sh
@@ -14,16 +19,29 @@ fi
 
 VERSION=$(cat /opt/$HOSTNAME/VERSION.txt)
 
-wget -O - -o /dev/null  \
-    --header="HOST:$HOSTNAME" \
-    ${URL:-http://127.0.0.1/} > /tmp/wget_$HOSTNAME.txt
-if grep -i "version: $VERSION" < /tmp/wget_$HOSTNAME.txt
-then 
-    rm -f /tmp/wget_$HOSTNAME.txt
-else
-    echo '***************** Problem!'
-    cat /tmp/wget_$HOSTNAME.txt
-    exit 32
-fi
+for HOSTNAME
+do (
+        echo "Inner check of $HOSTNAME"
+        source ${0%/*}/$HOSTNAME.sh
+        wget -O - -o /dev/null  \
+            --header="HOST:$HOSTNAME" \
+            ${URL:-http://127.0.0.1/} > /tmp/wget_$HOSTNAME.txt
+        if grep -i "version: $VERSION" < /tmp/wget_$HOSTNAME.txt
+        then 
+            if grep -q "${FRAGMENT:-}" < /tmp/wget_$HOSTNAME.txt
+            then
+                rm -f /tmp/wget_$HOSTNAME.txt
+            else
+                echo '***************** Problem (missing fragment)!'
+                cat /tmp/wget_$HOSTNAME.txt
+                exit 33
+            fi
+        else
+            echo '***************** Problem (wrong version)!'
+            cat /tmp/wget_$HOSTNAME.txt
+            exit 32
+        fi
+    )
+done
 
-# end of check-availability.sh
+# end of check-inner-availability.sh
